refactor(docs-gen): split extractor pipeline modules
